Refactor store mocks so they are usable in multiple packages
